feat(offers): anfragen@ inbound routing → extract → gated auto-send (Phase 2)
Wires the inbound pipeline end to end, behind the OFF kill switch:

- LLM provider: add extract_from_text (Protocol + Gemini via a shared
  _run_structured path + NullProvider), for structured extraction from an
  email body rather than a PDF/image.
- offer_extraction service (Celery-free, provider-injectable): pull
  art/object/units/desired_start/confidence from the inquiry body, write them
  onto the OfferInquiry + an llm_audit row; build_offer_request maps a
  post-extraction inquiry to an OfferGenerateRequest (None when fields are
  insufficient); auto_send_allowed gates on the kill switch + min confidence.
- inbound parser: capture mail.destination as `recipients`.
- webhook: route recipients whose local-part is `anfragen` to a new
  OfferInquiry + enqueue extraction (dedupe on message id), instead of the
  ticket flow. Local-part match so apex + inbound-subdomain both work.
- Celery task extract_offer_inquiry: extract → if confident + required fields
  + kill switch ON → generate offer + email it from anfragen@ (Resend
  from-override, PDF attached); else park as NEEDS_REVIEW. IGNORED for
  non-offer mail.
- confidence column typed Decimal; gating unit tests (kill switch defaults OFF).

Remaining: Admin review UI for NEEDS_REVIEW inquiries; runbook anfragen@ note.

+++ backend/app/models/anfrage.py
@@ -15,6 +15,7 @@ the `OfferInquiryStatus` StrEnum below is the canonical set used in code.
 import enum
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import Date, DateTime, Integer, Numeric, Text
 from sqlalchemy.dialects.postgresql import UUID
@@ -54,7 +55,7 @@ class OfferInquiry(OrganizationScopedMixin, TimestampMixin, Base):
     object_address: Mapped[str | None] = mapped_column(Text, nullable=True)
     units: Mapped[int | None] = mapped_column(Integer, nullable=True)
     desired_start: Mapped[date | None] = mapped_column(Date, nullable=True)
-    confidence: Mapped[float | None] = mapped_column(Numeric(4, 3), nullable=True)
+    confidence: Mapped[Decimal | None] = mapped_column(Numeric(4, 3), nullable=True)
     extraction_json: Mapped[str | None] = mapped_column(Text, nullable=True)
 
     # Outcome.
